Add CREATE VIEWS in db.sql, improve reliability
Add CREATE VIEW statements in db.sql so that they are automatically created
when running the script; also create an INDEX on the uic column of gares.
Otherwise, lookups take an awful lot of time.

Move the delay information after the destination name to unclutter the time
column.

Some other minor bugfixes here and there.