Enable autocompletion
Enable a fully functional autocompletion system.  This enables people to type
station names or codes without the hassle of selecting one in a list.

Backwards compatibility has been attempted, in that "ancient" browsers with
limited JavaScript support and iPhone users should be able to use the combobox
instead of the autocomp feature which would probably not work that well.