Add hack for trains calling at Ermont Eaubonne
This station seems to have two different UIC codes for some reason, one of
which is undocumented to make matters even worse.  These codes are used
interchangeably according to the train line.   Some trains cannot be found in
the database because of this.

This hack works around this glitch by trying again with the other station code
if the corresponding train cannot be found using the first one.