Improve cache efficiency; use Redis hashes
It has come to my attention that the data refreshed through the API is only
refreshed once a minute.  It is therefore useless to reqeuest for it more than
once per minute.

However, in order to keep the current average lag at 6 seconds instead of 30,
we don't cache the data for 1 minute by default.  We proceed as follows: if
the cached data changes while it was expired for less than 12 seconds, we keep
it for 60 seconds; otherwise we keep it for 12 seconds.